backend/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
email-validator==2.2.0
SQLAlchemy==2.0.39
joblib==1.4.2
pandas==2.2.3
scikit-learn==1.5.2
